==> v1_v2/work/1.1_extract_data_source.ipynb <==
%pip install sqlalchemy pymysql sshtunnel psycopg2-binary pytz

import pandas as pd
import environment as env
from utils import *

tunnel = open_ssh_tunnel(env.ssh_username, env.ssh_password, env.ssh_host, env.source_database_host, env.source_database_port)

connection = mysql_connect(env.source_database_username, env.source_database_password, env.source_database_host, env.source_database_port, env.source_database_name)

tables = run_query("SHOW TABLES;", connection)
tables

for table in tables["Tables_in_jcpmdb"]:
    run_query("SELECT * FROM " + table + ";", connection).to_pickle('./data_source/' + table + '.pickle')

connection.dispose()

tunnel.stop()
