==> v1_v2/work/2.1_transform_faq.ipynb <==
import pandas as pd
import environment as env
from utils import *

current_tenant = get_tenant(env.current_shopping)
current_tenant

faqs = pd.read_pickle('./data_source/faqs.pickle')
faqs

faqs = faqs.query("tenant_id == @current_tenant")
faqs = faqs.reset_index(drop=True)
faqs

faqs.loc[ :, 'tenant_id'] = env.current_shopping
faqs

faqs = faqs.rename(columns={"tenant_id": "shopping_id"})
faqs = faqs.drop(columns=['created_at', 'updated_at'])
faqs

faqs.to_pickle('./data/faq.pickle')
