==> v1_v2/work/2.2_transform_pricing.ipynb <==
import pandas as pd
import environment as env
from utils import *

current_tenant = get_tenant(env.current_shopping)
current_tenant

prices = pd.read_pickle('./data_source/prices.pickle')
prices

prices = prices.query("tenant_id == @current_tenant")
prices = prices.reset_index(drop=True)
prices

prices.loc[ :, 'tenant_id'] = env.current_shopping
prices

prices = prices.rename(columns={"tenant_id": "shopping_id"})
prices

prices["created_at"] = [ add_timezone_to_date(created_at) for created_at in prices["created_at"]]
prices["updated_at"] = [ add_timezone_to_date(updated_at) for updated_at in prices["updated_at"]]
prices

prices.to_pickle('./data/pricing.pickle')
