==> v1_v2/work/2.3_transform_customer.ipynb <==
import pandas as pd
import environment as env
from utils import *

current_tenant = get_tenant(env.current_shopping)
current_tenant

customers = pd.read_pickle('./data_source/customers.pickle')
customers

customers = customers.query("tenant_id == @current_tenant")
customers = customers.reset_index(drop=True)
customers

customers.loc[ :, 'tenant_id'] = env.current_shopping
customers

customers = customers.rename(columns={"tenant_id": "shopping_id"})
customers

customers["document_type"] = [doc.upper() if isinstance(doc, str) else doc for doc in customers["document_type"]]
customers

customers["created_at"] = [ add_timezone_to_date(created_at) for created_at in customers["created_at"]]
customers["updated_at"] = [ add_timezone_to_date(updated_at) for updated_at in customers["updated_at"]]
customers

customers.to_pickle('./data/customer.pickle')
