==> v1_v2/work/2.4_transform_card.ipynb <==
import pandas as pd
import environment as env
from utils import *

current_tenant = get_tenant(env.current_shopping)
current_tenant

customers = pd.read_pickle('./data/customer.pickle')
customers = customers.query("shopping_id == @env.current_shopping")
customers

cards = pd.read_pickle('./data_source/cards.pickle')
cards

cards = cards[cards.customer_id.isin(customers.id)]
cards = cards.reset_index(drop=True)
cards

cards.loc[ cards['type'] == 'CREDIT', 'type'] = 'CREDIT_CARD'
cards.loc[ cards['type'] == 'DEBIT', 'type'] = 'DEBIT_CARD'
cards

cards.loc[:, 'shopping_id'] = env.current_shopping
cards

cards["created_at"] = [ add_timezone_to_date(created_at) for created_at in cards["created_at"]]
cards["updated_at"] = [ add_timezone_to_date(updated_at) for updated_at in cards["updated_at"]]
cards

cards.to_pickle('./data/card.pickle')
