==> v1_v2/work/2.5_transform_ticket.ipynb <==
import pandas as pd
import environment as env
from utils import *

current_tenant = get_tenant(env.current_shopping)
current_tenant

customers = pd.read_pickle('./data/customer.pickle')
customers = customers.query("shopping_id == @env.current_shopping")
customers

tickets = pd.read_pickle('./data_source/tickets.pickle')
tickets

tickets = tickets[tickets.customer_id.isin(customers.id)]
tickets = tickets.reset_index(drop=True)
tickets

tickets.loc[:, 'shopping_id'] = env.current_shopping
tickets

tickets["checkin"] = [ add_timezone_to_date(checkin) for checkin in tickets["checkin"]]
tickets["checkout"] = [ add_timezone_to_date(checkout) for checkout in tickets["checkout"]]
tickets["next_checkout"] = [ add_timezone_to_date(next_checkout) for next_checkout in tickets["next_checkout"]]
tickets["created_at"] = [ add_timezone_to_date(created_at) for created_at in tickets["created_at"]]
tickets["updated_at"] = [ add_timezone_to_date(updated_at) for updated_at in tickets["updated_at"]]
tickets

tickets.to_pickle('./data/ticket.pickle')
