==> v1_v2/work/2.6_transform_transaction.ipynb <==
pip install pytz

import pandas as pd
import environment as env
from utils import *

current_tenant = get_tenant(env.current_shopping)
current_tenant

transactions = pd.read_pickle('./data_source/transactions.pickle')
transactions

transactions = transactions.query("tenant_id == @current_tenant")
transactions = transactions.reset_index(drop=True)
transactions

tickets = pd.read_pickle('./data/ticket.pickle')
tickets = tickets.query("shopping_id == @env.current_shopping")
tickets

transactions = transactions[transactions.ticket_id.isin(tickets.id)]
transactions = transactions.reset_index(drop=True)
transactions

transactions.loc[:, 'shopping_id'] = env.current_shopping
transactions = transactions.drop(columns=['tenant_id'])
# transactions = transactions.drop(columns=['external_id'])
transactions

transactions = transactions.rename(columns={"transaction_id": "card_transaction_id"})
transactions = transactions.rename(columns={"nsu": "card_transaction_nsu"})
transactions

transactions["checkin"] = [ add_timezone_to_date(checkin) for checkin in transactions["checkin"]]
transactions["checkout"] = [ add_timezone_to_date(checkout) for checkout in transactions["checkout"]]
transactions["next_checkout"] = [ add_timezone_to_date(next_checkout) for next_checkout in transactions["next_checkout"]]
transactions["created_at"] = [ add_timezone_to_date(created_at) for created_at in transactions["created_at"]]
transactions["updated_at"] = [ add_timezone_to_date(updated_at) for updated_at in transactions["updated_at"]]
transactions

transactions.to_pickle('./data/transaction.pickle')
