==> v1_v2/work/2.8_transform_advert.ipynb <==
import pandas as pd
import environment as env
from utils import *

current_tenant = get_tenant(env.current_shopping)
current_tenant

adverts = pd.read_pickle('./data_source/adverts.pickle')
adverts

adverts = adverts.query("tenant_id == @current_tenant")
adverts = adverts.reset_index(drop=True)
adverts

adverts.loc[ :, 'tenant_id'] = env.current_shopping
adverts

adverts = adverts.rename(columns={"tenant_id": "shopping_id"})
adverts

adverts["start_date"] = [ add_timezone_to_date(start_date) for start_date in adverts["start_date"]]
adverts["end_date"] = [ add_timezone_to_date(end_date) for end_date in adverts["end_date"]]
adverts["created_at"] = [ add_timezone_to_date(created_at) for created_at in adverts["created_at"]]
adverts["updated_at"] = [ add_timezone_to_date(updated_at) for updated_at in adverts["updated_at"]]
adverts

adverts.to_pickle('./data/advert.pickle')
