==> v1_v2/work/3.1_insert_faq.ipynb <==
%pip install sqlalchemy sshtunnel psycopg2-binary

import pandas as pd
import environment as env
from utils import *

tunnel = open_ssh_tunnel(env.ssh_username, env.ssh_password, env.ssh_host, env.target_database_host, env.target_database_port)

connection = postgres_connect(env.target_database_username, env.target_database_password, env.target_database_host, env.target_database_port, env.target_database_name)

target_faqs = pd.read_pickle('./data_target/faq.pickle')
target_faqs

string_target_faq_ids = [str(id) for id in target_faqs['id']]
string_target_faq_ids

faqs = pd.read_pickle('./data/faq.pickle')
faqs

faqs_to_insert = faqs[~faqs['id'].isin(string_target_faq_ids)]
faqs_to_insert

faqs_to_not_insert = faqs[faqs['id'].isin(string_target_faq_ids)]
faqs_to_not_insert

faqs_to_insert.to_sql('faq', connection, if_exists="append", index=False);

connection.dispose()

tunnel.stop()
