==> v1_v2/work/3.2_insert_pricing.ipynb <==
%pip install sqlalchemy sshtunnel psycopg2-binary

import pandas as pd
import environment as env
from utils import *

tunnel = open_ssh_tunnel(env.ssh_username, env.ssh_password, env.ssh_host, env.target_database_host, env.target_database_port)

connection = postgres_connect(env.target_database_username, env.target_database_password, env.target_database_host, env.target_database_port, env.target_database_name)

target_prices = pd.read_pickle('./data_target/pricing.pickle')
target_prices

string_target_price_ids = [str(id) for id in target_prices['id']]
string_target_price_ids

prices = pd.read_pickle('./data/pricing.pickle')
prices

prices_to_insert = prices[~prices['id'].isin(string_target_price_ids)]
prices_to_insert

prices_to_not_insert = prices[prices['id'].isin(string_target_price_ids)]
prices_to_not_insert

prices_to_insert.to_sql('pricing', connection, if_exists="append", index=False);

connection.dispose()

tunnel.stop()
