==> v1_v2/work/3.3_insert_customer.ipynb <==
%pip install sqlalchemy sshtunnel psycopg2-binary

import pandas as pd
import environment as env
from utils import *

tunnel = open_ssh_tunnel(env.ssh_username, env.ssh_password, env.ssh_host, env.target_database_host, env.target_database_port)

connection = postgres_connect(env.target_database_username, env.target_database_password, env.target_database_host, env.target_database_port, env.target_database_name)

target_customers = pd.read_pickle('./data_target/customer.pickle')
target_customers

string_target_customer_ids = [str(id) for id in target_customers['id']]
string_target_customer_ids

customers = pd.read_pickle('./data/customer.pickle')
customers

customers_to_insert = customers[~customers['id'].isin(string_target_customer_ids)]
customers_to_insert

customers_to_not_insert = customers[customers['id'].isin(string_target_customer_ids)]
customers_to_not_insert

customers_to_insert.to_sql('customer', connection, if_exists="append", index=False);

connection.dispose()

tunnel.stop()
