==> v1_v2/work/3.4_insert_card.ipynb <==
%pip install sqlalchemy sshtunnel psycopg2-binary

import pandas as pd
import environment as env
from utils import *

tunnel = open_ssh_tunnel(env.ssh_username, env.ssh_password, env.ssh_host, env.target_database_host, env.target_database_port)

connection = postgres_connect(env.target_database_username, env.target_database_password, env.target_database_host, env.target_database_port, env.target_database_name)

target_cards = pd.read_pickle('./data_target/card.pickle')
target_cards

string_target_card_ids = [str(id) for id in target_cards['id']]
string_target_card_ids

cards = pd.read_pickle('./data/card.pickle')
cards

cards_to_insert = cards[~cards['id'].isin(string_target_card_ids)]
cards_to_insert

cards_to_not_insert = cards[cards['id'].isin(string_target_card_ids)]
cards_to_not_insert

cards_to_insert.to_sql('card', connection, if_exists="append", index=False);

connection.dispose()

tunnel.stop()
