==> v1_v2/work/3.5_insert_ticket.ipynb <==
%pip install sqlalchemy sshtunnel psycopg2-binary

import pandas as pd
import environment as env
from utils import *

tunnel = open_ssh_tunnel(env.ssh_username, env.ssh_password, env.ssh_host, env.target_database_host, env.target_database_port)

connection = postgres_connect(env.target_database_username, env.target_database_password, env.target_database_host, env.target_database_port, env.target_database_name)

target_tickets = pd.read_pickle('./data_target/ticket.pickle')
target_tickets

string_target_ticket_ids = [str(id) for id in target_tickets['id']]
string_target_ticket_ids

tickets = pd.read_pickle('./data/ticket.pickle')
tickets

tickets_to_insert = tickets[~tickets['id'].isin(string_target_ticket_ids)]
tickets_to_insert

tickets_to_not_insert = tickets[tickets['id'].isin(string_target_ticket_ids)]
tickets_to_not_insert

tickets_to_insert.to_sql('ticket', connection, if_exists="append", index=False);

connection.dispose()

tunnel.stop()
