==> v1_v2/work/3.6_insert_transaction.ipynb <==
%pip install sqlalchemy sshtunnel psycopg2-binary

import pandas as pd
import environment as env
from utils import *

tunnel = open_ssh_tunnel(env.ssh_username, env.ssh_password, env.ssh_host, env.target_database_host, env.target_database_port)

connection = postgres_connect(env.target_database_username, env.target_database_password, env.target_database_host, env.target_database_port, env.target_database_name)

target_transactions = pd.read_pickle('./data_target/transaction.pickle')
target_transactions

string_target_transaction_ids = [str(id) for id in target_transactions['id']]
string_target_transaction_ids

transactions = pd.read_pickle('./data/transaction.pickle')
transactions

transactions_to_insert = transactions[~transactions['id'].isin(string_target_transaction_ids)]
transactions_to_insert

transactions_to_not_insert = transactions[transactions['id'].isin(string_target_transaction_ids)]
transactions_to_not_insert

transactions_to_insert.to_sql('transaction', connection, if_exists="append", index=False);

connection.dispose()

tunnel.stop()
